# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whisper-wrap engine API",
    "description": "HTTP/WebSocket/SSE API for the whisper-wrap transcription engine.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "3.0.0-beta.1"
  },
  "paths": {
...
    "/aux-models/download": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Start (or report an already-running) download. Returns immediately; poll\nGET /aux-models/download/{id}.",
        "operationId": "aux_models_download",
        "requestBody": {
          "description": "The auxiliary model id to download.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuxDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Download started/queued (ad-hoc JSON with a progress handle)."
          },
          "404": {
            "description": "Unknown auxiliary model id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Download setup failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/aux-models/download/{id}": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "aux_models_download_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Auxiliary model id whose download progress to read.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current download progress (ad-hoc JSON)."
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Cancel an in-flight aux download (mirror of models::cancel_download, keyed\nby aux id). The worker notices the flag between chunks.",
        "operationId": "aux_models_cancel_download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Auxiliary model id whose download to cancel.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled (ad-hoc JSON)."
          },
          "404": {
            "description": "No active download for that auxiliary model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/aux-models/{id}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Uninstall an aux model (delete its ONNX file from disk).",
        "operationId": "aux_models_delete_model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Auxiliary model id to uninstall.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Auxiliary model weights removed (ad-hoc JSON)."
          },
          "404": {
            "description": "Unknown auxiliary model id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Filesystem error removing the weights.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/models": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Installed + registered ASR models with active/loaded state (ad-hoc JSON)."
          },
          "500": {
            "description": "Registry read error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/download": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Start (or report an already-running) download of a model's ggml\nweights. Returns immediately; poll GET /models/download/{name}.",
        "operationId": "download",
        "requestBody": {
          "description": "The model name to download.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Download started/queued (ad-hoc JSON with a progress handle)."
          },
          "404": {
            "description": "Unknown model name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "A download for this model is already in progress.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Download setup failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/models/download/{name}": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "download_status",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Model name whose download progress to read.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current download progress (ad-hoc JSON)."
          },
          "404": {
            "description": "No active download for that model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Request cancellation of an in-flight download. The worker notices the\nflag between chunks, removes the partial file, then flips the job to\n\"cancelled\" — poll GET /models/download/{name} to observe it land.",
        "operationId": "cancel_download",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Model name whose download to cancel.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled (ad-hoc JSON)."
          },
          "404": {
            "description": "No active download for that model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/models/{name}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Uninstall a model's on-disk weights (the \"D\" in model CRUD). Refuses to\nremove the currently-loaded model — switch away first.",
        "operationId": "delete_model",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Model name to uninstall.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model weights removed (ad-hoc JSON)."
          },
          "404": {
            "description": "Unknown model name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Model weights missing or cannot be removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Filesystem error removing the weights.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The `{ \"detail\": string }` body every [`ApiError`] serializes to. `ApiError`\nitself is private and only implements `IntoResponse`, so it cannot derive\n`ToSchema`; this is the single reusable schema every fallible operation\nreferences for its non-200 error responses (rather than re-inlining the\nshape per handler).",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "Unsupported Content-Type: text/plain"
          }
        }
      },
      "AuxDownloadRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "DownloadRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "engine_token_bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "engine_token_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "engine_token"
      }
    }
  },
  "security": [
    {
      "engine_token_bearer": []
    },
    {
      "engine_token_cookie": []
    }
  ]
}